+++ NEWS.md
@@ -1,3 +1,8 @@
+# Quartet v1.1.0.9004 (development) 
+
+ - Remove errant name when comparing unnamed tree pairs.
+ - New function `SimilarityToReference()` to compare against reference topology.
+
 # Quartet v1.1.0.9003 (development) 
  
  - C++ implementation of `AllQuartets()`.